third_party/blink/web_tests/external/wpt/accelerometer/idlharness.https.window-expected.txt

This is a testharness.js-based test.
PASS idl_test setup
PASS idl_test validation
PASS Accelerometer interface: existence and properties of interface object
PASS Accelerometer interface object length
PASS Accelerometer interface object name
PASS Accelerometer interface: existence and properties of interface prototype object
PASS Accelerometer interface: existence and properties of interface prototype object's "constructor" property
PASS Accelerometer interface: existence and properties of interface prototype object's @@unscopables property
PASS Accelerometer interface: attribute x
PASS Accelerometer interface: attribute y
PASS Accelerometer interface: attribute z
PASS Accelerometer must be primary interface of new Accelerometer();
PASS Stringification of new Accelerometer();
PASS Accelerometer interface: new Accelerometer(); must inherit property "x" with the proper type
PASS Accelerometer interface: new Accelerometer(); must inherit property "y" with the proper type
PASS Accelerometer interface: new Accelerometer(); must inherit property "z" with the proper type
PASS LinearAccelerationSensor interface: existence and properties of interface object
PASS LinearAccelerationSensor interface object length
PASS LinearAccelerationSensor interface object name
PASS LinearAccelerationSensor interface: existence and properties of interface prototype object
PASS LinearAccelerationSensor interface: existence and properties of interface prototype object's "constructor" property
PASS LinearAccelerationSensor interface: existence and properties of interface prototype object's @@unscopables property
PASS LinearAccelerationSensor must be primary interface of new LinearAccelerationSensor();
PASS Stringification of new LinearAccelerationSensor();
PASS Accelerometer interface: new LinearAccelerationSensor(); must inherit property "x" with the proper type
PASS Accelerometer interface: new LinearAccelerationSensor(); must inherit property "y" with the proper type
PASS Accelerometer interface: new LinearAccelerationSensor(); must inherit property "z" with the proper type
FAIL GravitySensor interface: existence and properties of interface object assert_own_property: self does not have own property "GravitySensor" expected property "GravitySensor" missing
FAIL GravitySensor interface object length assert_own_property: self does not have own property "GravitySensor" expected property "GravitySensor" missing
FAIL GravitySensor interface object name assert_own_property: self does not have own property "GravitySensor" expected property "GravitySensor" missing
FAIL GravitySensor interface: existence and properties of interface prototype object assert_own_property: self does not have own property "GravitySensor" expected property "GravitySensor" missing
FAIL GravitySensor interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "GravitySensor" expected property "GravitySensor" missing
FAIL GravitySensor interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "GravitySensor" expected property "GravitySensor" missing
FAIL GravitySensor must be primary interface of new GravitySensor(); assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: GravitySensor is not defined"
FAIL Stringification of new GravitySensor(); assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: GravitySensor is not defined"
FAIL Accelerometer interface: new GravitySensor(); must inherit property "x" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: GravitySensor is not defined"
FAIL Accelerometer interface: new GravitySensor(); must inherit property "y" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: GravitySensor is not defined"
FAIL Accelerometer interface: new GravitySensor(); must inherit property "z" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: GravitySensor is not defined"
Harness: the test ran to completion.
